test(Activity) Adicionado testes para a formatacao
