+++ solutions/devsprint-cecilia-saraiva-2/FinanceApp/Service/Entities/HomeData.swift
@@ -14,17 +14,3 @@ struct HomeData: Decodable {
     let spending: Float
     let activity: [Activity]
 }
-
-struct Activity: Decodable {
-    let name: String
-    let price: Float
-    let time: String
-}
-
-extension Activity {
-
-    func formattedInfo() -> String {
-
-        return "$\(price) • \(time)"
-    }
-}
